Ensure unique junctions don't result in self-edges
When uniquifying junction nodes, we previously ignored the fact that
they would result in self-edges in the pixel graph, because multiple
adjacent pixels would have the same label/ID. Now, by the judicious
modification of an if statement, we avoid creating these edges
altogether. An informal benchmark shows that this does not affect
performance of graph build time.