Improved error reporting and tests for prune_paths() methods
Closes #206

The error arises during usage of the `csr.Skeleton.prune_paths()` which takes a list of indices from
`csr.summarize()` which are to be pruned. If the index is outside of the range of rows in the data frame
a `ValueError` is thrown.

It was first highlighted during early development of pruning iteratively but can arise any time in inappropriate value
that is greater than the number of paths in a skeleton is passed to `prune_path()` method.

A `try: ... except: ...` is introduced to capture this and report, hopefully informatively, what the problem is and how
to correct it. Tests are introduced for this method, parametrized to test removing different types of paths along with
tests of two exceptions that can arise.