Networkx <--> Skeleton development and tests
I was looking to convert NetworkX objects back to Skeletons and stumbled on the `nx.to_numpy_array()` function and the
related `nx.from_numpy_array()`.

Since `csr.skeleton_image` are typically Numpy Arrays I figured this would be a convenient and efficient method of
converting between the two formats so have attempted to replace `skeleton_to_nx()` with this.

I had written some tests for `skeleton_to_nx()` first before making undertaking refactoring.

I found that `nx.from_numpy_array()` doesn't handle non-square arrays, ok not a problem we can pad them with
zeros (wrote a small function to do this along with tests), but on applying `nx.from_numpy_array()` to these padded
arrays the tests I wrote to originally failed, which surprised me as, in my naivety, I wouldn't expect padding with
zero's to adjacency graphs to affect the resulting network representation in terms of the number of nodes and the number
of edges.