Adds array_to_nx() and nx_to_array() functions
Introduces conversion of Numpy arrays (_not_ adjacency matrices) to NetworkX graphs.

This is done by assuming that each element of a 2-D boolean array[1] is a node connected by edges, we look at what
values the neighbours are and > 0 the two points form an edge which is added to a NetworkX graph.

The shape of the original Numpy array is saved within the `Networkx.graph` object which makes
reconstructing the original Numpy array straight-forward (as we may be pruning some of the original co-ordinates may be
removed).

Thanks to @MaxGamill-Sheffield for guiding me to this approach and as noted a solution found on SO for how to find the
co-ordinates of neighbouring cells.