Reconstruct Skeleton from stored path_coordinates()
Saves the `.path_coordinates()` to the edge properties of a NetworkX graph, using the `row.Index` _and_ the
`node_id_src` / `node_id_dst` to define the nodes. This is required because in some instances there are loops with the
same `src` and `dst` and if only these are used we would lose information, including the `row.Index` as a proxy for the
graph segment avoids this.

As the `.path_coordinates()` are saved as `edge` properties in NetworkX object we can now reconstruct the original Numpy
array based on this information and convert to a `Skeleton`.

Tests are included for a range of sample graphs and exceptions are raised (and tested) when...

+ Wrong type of object is passed into `nx_to_skeleton()`.
+ The NetworkX object is missing `edge` properties completely.
+ The `edge` properties are co-ordinates outside of the original images dimensions.

There may be more edge cases that we could/should test for.

I found problems with `tinyline` as it was a 1-D "image" and so have made this a 2-D "image" with a single row and
updated `test_line()` appropriately.