Don't use full version info for deploy action (#243)
Another attempt to fix deployment with
JamesIves/github-pages-deploy-action.

But to be honest I am just fumbling around in the dark here. The only
information I can find about "Missing download info" in GHA is this
discussion:

https://github.com/orgs/community/discussions/152695

which relates to a version of actions/upload-artifact that was no
longer available. That doesn't seem to be the case for
JamesIves/github-pages-deploy-action, which currently shows the 4.7.3
tag on the repo.