New API: PathGraph dataclass (#246)
The Skeleton class is a bit of a hodgepodge of data, generated data,
things that should be properties, manually cached properties, and
functions that should not be part of the class at all. Thanks to the
helpful prodding of @kevinyamauchi, this PR attempts to simplify the
concepts and data structures in skan into a simple class that can be
created without an image — since the path graph, not the image, is the
core of the computational abilities of skan.

This is still a work in progress but should already be useful: if you
have a graph as a `scipy.sparse.csr_array` (note: `csr_array`, not the
deprecated `csr_matrix` used by skan so far) generated through your own
means, you can make a "Skeleton" as follows:

```python
from skan.csr import PathGraph, Skeleton, summarize

g = PathGraph.from_graph(
        node_coordinates=coordinates,  # (n, ndim) NumPy array
        graph=graph,  # scipy.sparse.csr_array
        )

s = Skeleton.from_path_graph(g)
summary = summarize(s, separator='_')
```

Still to do in this PR:

- allow making PathGraphs from Skeletons
- allow `summarize` to take in PathGraphs directly

In the future, we might want to deprecate Skeleton altogether, but I'm
happy to do this over a long time.

@kevinyamauchi, I'm curious what you think about `paths` being a data
attribute, even though it is generated. I think it's expensive enough to
compute that we want it to be data and serializable. But it kinda breaks
the dataclass paradigm a little bit.

@DragaDoncila, you should be able to pull down this branch and use it on
your networkx tracking graphs after exporting them to csr arrays (I'm
pretty sure that's built-in to nx).